Перенес функции даты в отдельный файл
